api/requirements.txt
Flask>=1.1.2
serverless-wsgi>=2.0.2
Flask-Migrate==2.1.1
Flask-Script==2.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.15